Reduce file size support to 4 GB.

+++ src/fileindex.h
@@ -5,4 +5,4 @@ typedef uint fileref;
 extern void FileIndexInit(void);
 extern nonnull fileref FileIndexAdd(const char* filename);
 extern pure const byte* FileIndexGetContents(fileref file);
-extern pure ulong FileIndexGetSize(fileref file);
+extern pure uint FileIndexGetSize(fileref file);
